+++ spec/storey/duplicator_spec.rb
@@ -4,18 +4,19 @@ describe Storey::Duplicator do
 
   describe '#perform!' do
     context "when the dump is a failure" do
-      it 'should raise an error' do
+      it 'raises an error' do
         duplicator = described_class.new('non-existent-will-fail-dump', 'new')
-        expect {
-          duplicator.perform!
-        }.to raise_error(
-          Storey::StoreyError,
-          "There seems to have been a problem dumping `non-existent-will-fail-dump` to make a copy of it into `new`"
-        )
+        expected_msg = [
+          "Problem dumping `non-existent-will-fail-dump` to make a copy of it",
+          "into `new`: pg_dump: no matching schemas were found",
+        ].join(" ")
+        expect { duplicator.perform! }.
+          to raise_error(Storey::StoreyError).
+          with_message(/#{expected_msg}/)
       end
     end
 
-    it 'should remove the target and source sql files after work' do
+    it 'removes the target and source sql files after work' do
       Storey.create 'boo'
       duplicator = described_class.new('boo', 'ya')
       duplicator.perform!