+++ spec/lib/storey_spec.rb
@@ -13,4 +13,19 @@ RSpec.describe Storey do
     end
   end
 
+  describe ".switch" do
+    it "does not cache between switches" do
+      Storey.create("s1") do
+        2.times {|n| Post.create(name: n.to_s) }
+      end
+
+      Storey.create("s2") do
+        3.times {|n| Post.create(name: n.to_s) }
+      end
+
+      Storey.switch("s1") { expect(Post.count).to eq 2 }
+      Storey.switch("s2") { expect(Post.count).to eq 3 }
+    end
+  end
+
 end