+++ spec/spec_helper.rb
@@ -14,6 +14,12 @@ require 'pry'
 RSpec.configure do |config|
   config.order = 'random'
 
+  config.before(:suite) do
+    # Enable query cache so we can catch unexpected behaviour with AR
+    # caching
+    ActiveRecord::Base.connection.enable_query_cache!
+  end
+
   config.before(:each) do
     # We don't want configuration to leak into other tests
     Storey.reload_config!