+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.2] - 2018-08-14
+### Fixed
+- Instances when queries are cached between switches
+
 ## [2.1.1] - 2018-07-04
 ### Fixed
 - Fix copying of schemas in when ActiveRecord caches queries